requirements.txt
opencv-python
pandas
numpy
scikit-learn
matplotlib
streamlit
seaborn
plotly
Flask
Pillow
python-dotenv
requests
Werkzeug